requirements.txt
pyspark
kafka-python
yfinance
pandas
jupyterlab
matplotlib